move data load functions to utils
