+++ src/markermap/utils.py
@@ -11,6 +11,7 @@ from torch.nn import functional as F
 from pathlib import Path
 import sys
 import os
+import gc
 import contextlib
 import queue
 from torch.utils.data import DataLoader, random_split
@@ -2139,6 +2140,9 @@ def quick_model_summary(model, train_data, test_data, threshold, batch_size):
     print("# Non Sparse in Orig test")
     print(torch.sum(test_data[0,:] != 0))
 
+
+## Data Loading and Preprocessing Helper Functions
+
 # X can be a numpy array 
 def process_data(X, Y, filter_data = False):
     adata = sc.AnnData(X)
@@ -2183,6 +2187,202 @@ def parse_adata(adata):
     y = encoder.transform(labels)
     return X, y, encoder
 
+def get_zeisel(file_path):
+    """
+    Get the zeisel data located a file path
+    args:
+        file_path (string): location of zeisel.h5ad file
+    returns: X data array, y transformed labels, encoder that transformed the labels
+    """
+    adata = sc.read_h5ad(file_path)
+    adata.obs['names']=adata.obs['names0']
+    adata.obs['annotation'] = adata.obs['names0']
+
+    return parse_adata(adata)
+
+def get_paul(housekeeping_bone_marrow_path, house_keeping_HSC_path):
+    """
+    Get the paul data from scanpy and remove the housekeeping genes from the two provided files
+    args:
+        housekeeping_bone_marrow_path (string): location of housekeeping genes bone marrow file
+        house_keeping_HSC_path (string): location of housekeeping genes HSC file
+    returns: X data array, y transformed labels, encoder that transformed the labels
+    """
+    adata = sc.datasets.paul15()
+    sm = SmashPyWrapper()
+    sm.data_preparation(adata)
+    adata = sm.remove_general_genes(adata)
+    adata = sm.remove_housekeepingenes(adata, path=[housekeeping_bone_marrow_path])
+    adata = sm.remove_housekeepingenes(adata, path=[house_keeping_HSC_path])
+    dict_annotation = {}
+
+    dict_annotation['1Ery']='Ery'
+    dict_annotation['2Ery']='Ery'
+    dict_annotation['3Ery']='Ery'
+    dict_annotation['4Ery']='Ery'
+    dict_annotation['5Ery']='Ery'
+    dict_annotation['6Ery']='Ery'
+    dict_annotation['7MEP']='MEP'
+    dict_annotation['8Mk']='Mk'
+    dict_annotation['9GMP']='GMP'
+    dict_annotation['10GMP']='GMP'
+    dict_annotation['11DC']='DC'
+    dict_annotation['12Baso']='Baso'
+    dict_annotation['13Baso']='Baso'
+    dict_annotation['14Mo']='Mo'
+    dict_annotation['15Mo']='Mo'
+    dict_annotation['16Neu']='Neu'
+    dict_annotation['17Neu']='Neu'
+    dict_annotation['18Eos']='Eos'
+    dict_annotation['19Lymph']='Lymph'
+
+    annotation = []
+    for celltype in adata.obs['paul15_clusters'].tolist():
+        annotation.append(dict_annotation[celltype])
+
+    adata.obs['annotation'] = annotation
+    adata.obs['annotation'] = adata.obs['annotation'].astype('category')
+
+    return parse_adata(adata)
+
+def get_citeseq(file_path):
+    """
+    Get the CITE_seq data located a file path
+    args:
+        file_path (string): location of CITEseq.h5ad file
+    returns: X data array, y transformed labels, encoder that transformed the labels
+    """
+    adata = sc.read_h5ad(file_path)
+    adata.obs['annotation'] = adata.obs['names']
+    return parse_adata(adata)
+
+def relabel_mouse_labels(label):
+    if isinstance(label, str):
+        return label.split('_')[0]
+    else:
+        return label
+
+def remove_features_pct(adata, group_by=None, pct=0.3):
+    if group_by is None:
+        print("select a group_by in .obs")
+        return
+    if group_by not in adata.obs.columns:
+        print("group_by must be in .obs")
+        return
+
+
+    list_keep_genes = []
+
+    df = pd.DataFrame(data=False, index=adata.var.index.tolist(), columns=adata.obs[group_by].cat.categories)
+    for g in adata.obs[group_by].cat.categories:
+        reduced = adata[adata.obs[group_by]==g]
+        boolean, values = sc.pp.filter_genes(reduced, min_cells = reduced.n_obs*pct, inplace=False)
+        df[g] = boolean
+    dfT = df.T
+    for g in dfT.columns:
+        if True in dfT[g].tolist():
+            list_keep_genes.append(True)
+        else:
+            list_keep_genes.append(False)
+
+    adata.var["general"] = list_keep_genes
+
+    adata = adata[:, adata.var["general"]]
+
+    return adata
+
+def remove_features_pct_2groups(adata, group_by=None, pct1=0.9, pct2=0.5):
+    if group_by is None:
+        print("select a group_by in .obs")
+        return
+    if group_by not in adata.obs.columns:
+        print("group_by must be in .obs")
+        return
+
+
+    list_keep_genes = []
+
+    df = pd.DataFrame(data=False, index=adata.var.index.tolist(), columns=adata.obs[group_by].cat.categories)
+    for g in adata.obs[group_by].cat.categories:
+        reduced = adata[adata.obs[group_by]==g]
+        boolean, values = sc.pp.filter_genes(reduced, min_cells = reduced.n_obs*(pct1), inplace=False)
+        df[g] = boolean
+    dfT = df.T
+    for g in dfT.columns:
+        if (sum(dfT[g].tolist())/len(dfT[g].tolist())) >= pct2:
+            list_keep_genes.append(False)
+        else:
+            list_keep_genes.append(True)
+
+    adata.var["general"] = list_keep_genes
+    adata = adata[:, adata.var["general"]]
+    return adata
+
+def get_mouse_brain(mouse_brain_path, mouse_brain_labels_path):
+    """
+    Get the mouse brain data and remove outliers and perform normalization. Some of the decisions in this function are
+    judgement calls, so users should inspect and make their own decisions.
+    args:
+        mouse_brain_path (string): location of mouse_brain_all_cells.h5ad file
+        mouse_brain_labels_path (string): location of cells annotations for mouse_brain data
+    returns: X data array, y transformed labels, encoder that transformed the labels
+    """
+    adata_snrna_raw = anndata.read_h5ad(mouse_brain_path)
+    del adata_snrna_raw.raw
+    adata_snrna_raw = adata_snrna_raw
+    adata_snrna_raw.X = adata_snrna_raw.X.toarray()
+    ## Cell type annotations
+    labels = pd.read_csv(mouse_brain_labels_path, index_col=0)
+    labels['annotation'] = labels['annotation_1'].apply(lambda x: relabel_mouse_labels(x))
+    labels = labels[['annotation']]
+    labels = labels.reindex(index=adata_snrna_raw.obs_names)
+    adata_snrna_raw.obs[labels.columns] = labels
+    adata_snrna_raw = adata_snrna_raw[~adata_snrna_raw.obs['annotation'].isna(), :]
+    adata_snrna_raw = adata_snrna_raw[adata_snrna_raw.obs["annotation"]!='Unk']
+    adata_snrna_raw.obs['annotation'] = adata_snrna_raw.obs['annotation'].astype('category')
+
+    # doing this cuz smash was a lot of data
+    #https://cell2location.readthedocs.io/en/latest/notebooks/cell2location_estimating_signatures.html
+    #preprocess_like_cell_location
+    sc.pp.filter_cells(adata_snrna_raw, min_genes=1)
+    print(adata_snrna_raw.shape)
+    sc.pp.filter_genes(adata_snrna_raw, min_cells=1)
+    print(adata_snrna_raw.shape)
+
+    gc.collect()
+    # calculate the mean of each gene across non-zero cells
+    adata_snrna_raw.var['n_cells'] = (adata_snrna_raw.X > 0).sum(0)
+    adata_snrna_raw.var['nonz_mean'] = adata_snrna_raw.X.sum(0) / adata_snrna_raw.var['n_cells']
+
+    nonz_mean_cutoff = np.log10(1.12) # cut off for expression in non-zero cells
+    cell_count_cutoff = np.log10(adata_snrna_raw.shape[0] * 0.0005) # cut off percentage for cells with higher expression
+    cell_count_cutoff2 = np.log10(adata_snrna_raw.shape[0] * 0.03)# cut off percentage for cells with small expression
+
+
+    adata_snrna_raw[:,(np.array(np.log10(adata_snrna_raw.var['nonz_mean']) > nonz_mean_cutoff)
+        | np.array(np.log10(adata_snrna_raw.var['n_cells']) > cell_count_cutoff2))
+        & np.array(np.log10(adata_snrna_raw.var['n_cells']) > cell_count_cutoff)].shape
+
+    # select genes based on mean expression in non-zero cells
+    adata_snrna_raw = adata_snrna_raw[:,(np.array(np.log10(adata_snrna_raw.var['nonz_mean']) > nonz_mean_cutoff)
+        | np.array(np.log10(adata_snrna_raw.var['n_cells']) > cell_count_cutoff2))
+        & np.array(np.log10(adata_snrna_raw.var['n_cells']) > cell_count_cutoff)
+        & np.array(~adata_snrna_raw.var['SYMBOL'].isna())]
+
+    gc.collect()
+    adata_snrna_raw.raw = adata_snrna_raw
+    adata_snrna_raw.X = adata_snrna_raw.raw.X.copy()
+    del adata_snrna_raw.raw
+    gc.collect()
+    adata_snrna_raw = remove_features_pct(adata_snrna_raw, group_by="annotation", pct=0.3)
+    gc.collect()
+    adata_snrna_raw = remove_features_pct_2groups(adata_snrna_raw, group_by="annotation", pct1=0.75, pct2=0.5)
+
+    sc.pp.normalize_per_cell(adata_snrna_raw, counts_per_cell_after=1e4)
+    sc.pp.log1p(adata_snrna_raw)
+    sc.pp.scale(adata_snrna_raw, max_value=10)
+
+    return parse_adata(adata_snrna_raw)
 
 def split_data_into_dataloaders_no_test(X, Y, train_size, batch_size = 64, num_workers = 0, seed = None):
     """
